PHPstan level 5

+++ src/Service/Command/ConsoleHelper.php
@@ -231,7 +231,7 @@ class ConsoleHelper
     public static function progressBar(int $current, int $total, int $width = 50): void
     {
         $percent = ($current / $total);
-        $filled = round($percent * $width);
+        $filled = (int)round($percent * $width);
         $bar = str_repeat('█', $filled).str_repeat('░', $width - $filled);
         $percentDisplay = str_pad(round($percent * 100).'%', 4, ' ', STR_PAD_LEFT);
         echo "\r".self::color("Progress: [{$bar}] {$percentDisplay}", '36');
@@ -243,7 +243,7 @@ class ConsoleHelper
     /**
      * Retourne une icône en fonction du type de fichier.
      *
-     * @param string \$path
+     * @param string $path
      */
     public static function fileIcon(string $path): string
     {
@@ -268,7 +268,7 @@ class ConsoleHelper
     /**
      * Retourne une couleur ANSI suggérée pour un type de fichier.
      *
-     * @param string \$path
+     * @param string $path
      */
     public static function fileColor(string $path): string
     {
